Add initial WindowedCells implementation
