+++ package.json
@@ -171,6 +171,8 @@
     "@types/react-table": "^6.7.21",
     "@types/react-test-renderer": "^16.0.3",
     "@types/react-timeago": "^4.1.0",
+    "@types/react-virtualized-auto-sizer": "^1.0.0",
+    "@types/react-window": "^1.8.1",
     "@types/redux-immutable": "^4.0.0",
     "@types/redux-logger": "^3.0.6",
     "@types/semver": "^6.0.0",
@@ -241,7 +243,7 @@
     "webfontloader": "^1.6.28",
     "webpack": "^4.28.3",
     "webpack-cli": "^3.3.7",
-    "webpack-dev-server": "^3.1.14",
+    "webpack-dev-server": "3.7.1",
     "webpack-merge": "^4.2.1",
     "yargs": "^13.0.0"
   },
